Add new tacky instructions

+++ main.c
@@ -1,5 +1,5 @@
 int main(void) {
     /* "Unsigned double" is not a valid type specifier */
-    unsigned double d = 10.0;
+    double d = 10.0;
     return 0;
 }
